feat: organize
